refactor: rename maelvls/quote -> maelvls/users-grpc
The forst name, 'quote', was my first idea of a microservice. But I
could not find any cool things around 'quotes'. Then I though: maybe a
todo microservice? Or may be an addressbook?

Then it was almost too late and 3 days before the deadline, I chose to
go with users.